README.md
The Final Year Project is required by Heritage Institute of Technology, Kolkata as partial fulfillment of the bachelor degree in Computer Science and Engineering. Throughout the development of this project, the student about to graduate demonstrates the knowledge acquired during his four years as an undergraduate student. The development of this project required the use of most of the learned computer science skills: Database Design, Android Development and Software Engineering.

This project has basically 2 modules. The first module includes Customer module and second module includes Admin module

There are lot more feature I'm going to add in future.